Use dm-core/spec/setup in specs
Specs pass on all official adapters and on all
supported rubies. The in_memory and yaml adapters
don't support all operations though. Those are
marked pending in the specs, when the specs are
run with any of those adapters.

Instructions on how to run the specs can be found
in the Gemfile